+++ sweat/__init__.py
@@ -5,6 +5,6 @@ from .io.gpx import read_gpx
 from .io.generic import read_dir, read_file
 from .io.strava import read_strava
 from .io.tcx import read_tcx
-from .pdm.regressors import CriticalPowerRegressor
+from .pdm.regressors import PowerDurationRegressor
 from .pdm.w_prime_balance import w_prime_balance
 from .utils import array_1d_to_2d